integrations/llamaindex/python/requirements.txt
llama-index-core>=0.11.0,<0.12.0
llama-index-llms-openai>=0.2.0
llama-index-embeddings-openai>=0.2.0
numpy>=1.26.0
